+++ apps/api/src/Jobby.Application/Features/ContactFeatures/Commands/Delete/DeleteContactCommandHandler.cs
@@ -1,12 +1,13 @@
 ﻿using Jobby.Application.Abstractions.Specification;
 using Jobby.Application.Interfaces.Services;
+using Jobby.Application.Responses.Common;
 using Jobby.Application.Services;
 using Jobby.Domain.Entities;
 using MediatR;
 
 namespace Jobby.Application.Features.ContactFeatures.Commands.Delete;
 
-internal sealed class DeleteContactCommandHandler : IRequestHandler<DeleteContactCommand, Unit>
+internal sealed class DeleteContactCommandHandler : IRequestHandler<DeleteContactCommand, BaseResult<DeleteContactResponse, DeleteContactOutcomes>>
 {
     private readonly IRepository<Contact> _contactRepository;
     private readonly string _userId;
@@ -19,14 +20,34 @@ internal sealed class DeleteContactCommandHandler : IRequestHandler<DeleteContac
         _userId = userService.UserId();
     }
 
-    public async Task<Unit> Handle(DeleteContactCommand request, CancellationToken cancellationToken)
+    public async Task<BaseResult<DeleteContactResponse, DeleteContactOutcomes>> Handle(DeleteContactCommand request, CancellationToken cancellationToken)
     {
-        Contact contactToDelete = await ResourceProvider<Contact>
+        var contactResourceResult = await ResourceProvider<Contact>
             .GetById(_contactRepository.GetByIdAsync)
             .Check(_userId, request.ContactId, cancellationToken);
 
+        if (!contactResourceResult.IsSuccess)
+        {
+            return new BaseResult<DeleteContactResponse, DeleteContactOutcomes>(
+                IsSuccess: false,
+                Outcome: contactResourceResult.Outcome switch
+                {
+                    Outcome.Unauthorised => DeleteContactOutcomes.UnauthorizedContactAccess,
+                    Outcome.NotFound => DeleteContactOutcomes.UnknownContact,
+                    _ => DeleteContactOutcomes.UnknownError
+                },
+                ErrorMessage: contactResourceResult.ErrorMessage
+            );
+        }
+        
+        var contactToDelete = contactResourceResult.Response;
+
         await _contactRepository.DeleteAsync(contactToDelete, cancellationToken);
 
-        return Unit.Value;
+        return new BaseResult<DeleteContactResponse, DeleteContactOutcomes>(
+            IsSuccess: true,
+            Outcome: DeleteContactOutcomes.ContactDeleted,
+            Response: new DeleteContactResponse()
+        );
     }
 }
